# services/gateway/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcommerceGo API Gateway",
    "description": "Unified entry point for all EcommerceGo microservices. All requests require a JWT Bearer token in the Authorization header unless noted otherwise. Prices are always in cents (int64).",
    "version": "1.0.0",
    "contact": {
      "name": "EcommerceGo",
      "url": "https://github.com/utafrali/EcommerceGo"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (API Gateway)"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "tags": ["orders"],
        "summary": "List user orders",
        "operationId": "listOrders",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Paginated order list" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["orders"],
        "summary": "Create order",
        "operationId": "createOrder",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrderRequest" } } } },
        "responses": {
          "201": { "description": "Order created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": ["orders"],
        "summary": "Get order by ID",
        "operationId": "getOrder",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "responses": {
          "200": { "description": "Order details" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/orders/{id}/status": {
      "put": {
        "tags": ["orders"],
        "summary": "Update order status",
        "operationId": "updateOrderStatus",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateOrderStatusRequest" } } }
        },
        "responses": {
          "200": { "description": "Order status updated" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/v1/payments": {
      "get": {
        "tags": ["payments"],
        "summary": "List user payments",
        "operationId": "listPayments",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Payment list" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/payments/{id}": {
      "get": {
        "tags": ["payments"],
        "summary": "Get payment by ID",
        "operationId": "getPayment",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "responses": {
          "200": { "description": "Payment details" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/payments/{id}/refund": {
      "post": {
        "tags": ["payments"],
        "summary": "Refund payment",
        "operationId": "refundPayment",
        "parameters": [ { "$ref": "#/components/parameters/IDParam" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefundRequest" } } } },
        "responses": {
          "200": { "description": "Refund initiated" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login or /api/v1/auth/register"
      }
    },
    "parameters": {
      "IDParam": {
        "name": "id",
        "in": "path",
        "description": "Resource UUID",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "UNAUTHORIZED" },
          "message": { "type": "string", "example": "missing or invalid token" },
          "request_id": { "type": "string", "example": "req-abc123" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "required": ["user_id", "items"],
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["product_id", "quantity", "unit_price"],
              "properties": {
                "product_id": { "type": "string", "format": "uuid" },
                "variant_id": { "type": "string", "format": "uuid" },
                "quantity": { "type": "integer", "minimum": 1 },
                "unit_price": { "type": "integer", "format": "int64" }
              }
            }
          },
          "shipping_address_id": { "type": "string", "format": "uuid" },
          "coupon_code": { "type": "string" }
        }
      },
      "UpdateOrderStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "enum": ["pending", "confirmed", "shipped", "delivered", "cancelled"] }
        }
      },
...
      "RefundRequest": {
        "type": "object",
        "required": ["amount"],
        "properties": {
          "amount": { "type": "integer", "format": "int64", "minimum": 1, "description": "Refund amount in cents" },
          "reason": { "type": "string" }
        }
      },
...
    }
  }
}